Release 3.10.0 (tracking kernel release 3.10.0) incorporates the following changes: - removed autotools-generated files from the repo - more cleanup to reflect the move to git - new BUILD.md document describes the vagrant-based development environment - upgraded to CentOS 7.1 build server - updated release notes (in NEWS) - fixed some missing includes and lint-type errors

+++ aprsshm.c
@@ -68,8 +68,8 @@ static struct pktseg *pktseg = NULL;
 static char *infofile = NULL;
 
 /* convert relative ptr to absolute & vice-versa */
-#define RELABS(x) ((struct pkt *)(((unsigned int)(x))+(unsigned int)pktseg))
-#define ABSREL(x) ((struct pkt *)(((unsigned int)(x))-(unsigned int)pktseg))
+#define RELABS(x) ((struct pkt *)(((intptr_t)(x))+(intptr_t)pktseg))
+#define ABSREL(x) ((struct pkt *)(((intptr_t)(x))-(intptr_t)pktseg))
 #define PKTSIZE (sizeof(struct pkt))
 
 int
